backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
celery==5.4.0
redis==5.1.1
python-multipart==0.0.10
httpx==0.27.2
python-docx==1.1.2
pypdf==5.0.1
qrcode==7.4.2
pytest==8.3.3
alembic==1.13.3
